Reuse code
